added readAllIngredients

+++ src/main/java/demo/data/api/Ingredients.java
@@ -3,7 +3,7 @@ package demo.data.api;
 public interface Ingredients {
     String getName();
     void setName(String name);
-    int getQuantity();
-    void setQuantity(int quantity);
+    float getQuantity();
+    void setQuantity(float quantity);
 
 }
